src/backend/requirements.txt
aiohttp==3.9.4rc0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
azure-communication-callautomation==1.1.0
azure-communication-identity==1.5.0
azure-core==1.30.1
azure-eventgrid==4.17.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dapr==1.13.0
dapr-ext-fastapi==1.12.0
decorator==5.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
frozenlist==1.4.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
idna==3.6
isodate==0.6.1
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
protobuf==4.25.3
py==1.11.0
pydantic==2.6.3
pydantic-core==2.16.3
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
yarl==1.9.4
